renamed files, moved my changes into merged GUI
